backend/requirements.txt
amqp==2.2.2
aniso8601==7.0.0
attrs==17.4.0
autopep8==1.5
billiard==3.5.0.3
botocore==1.13.46
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-filter==2.2.0
django-log-request-id==1.3.2
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-cors-headers==3.4.0    
docutils==0.14
graphene==2.1.8
graphene-django==2.8.1
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
model-mommy==1.5.1
more-itertools==4.1.0
openapi-codec==1.3.2
pbr==4.0.1
pika==0.11.2
pluggy==0.6.0
promise==2.3
psycopg2==2.7.4
py==1.5.3
pycodestyle==2.5.0
Pygments==2.2.0
PyJWT==1.6.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.20.0
responses==0.9.0
Rx==1.6.1
s3transfer==0.1.13
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
